+++ web/vueapp/src/app.vue
@@ -16,7 +16,7 @@
   import store from './store';
   import { router, i18n } from './bootstrap';
   import accountService from './services/account';
-  import checkCookie from './utils/cookieNotification';
+  import { checkCookie, checkLocalStorageAvailable } from './utils/cookieNotification';
   import checkEmailVerification from './utils/emailVerificationNotification';
 
   export default {
@@ -44,6 +44,7 @@
       if (this.$store.state.auth.authenticated) {
         accountService.find();
       }
+      checkLocalStorageAvailable();
       setTimeout(checkCookie, 2000);
       setTimeout(checkEmailVerification, 10000);
     },