+++ web/vueapp/src/services/storage/getItem.js
@@ -1,0 +1,7 @@
+export default (key) => {
+  try {
+    return localStorage.getItem(key);
+  } catch (e) {
+    return undefined;
+  }
+};