+++ web/vueapp/src/services/storage/hasItem.js
@@ -1,0 +1,9 @@
+import getItem from './getItem';
+
+export default (key) => {
+  try {
+    return !!getItem(key);
+  } catch (e) {
+    return false;
+  }
+};