+++ web/vueapp/src/services/storage/index.js
@@ -1,0 +1,13 @@
+import getItem from './getItem';
+import hasItem from './hasItem';
+import setItem from './setItem';
+import removeItem from './removeItem';
+import isAvailable from './isAvailable';
+
+export default {
+  getItem,
+  hasItem,
+  setItem,
+  removeItem,
+  isAvailable,
+};