+++ web/vueapp/src/services/storage/isAvailable.js
@@ -1,0 +1,9 @@
+export default () => {
+  try {
+    localStorage.setItem('test', 'data');
+    localStorage.removeItem('test');
+    return true;
+  } catch (e) {
+    return false;
+  }
+};