+++ web/vueapp/src/services/storage/removeItem.js
@@ -1,0 +1,6 @@
+export default (key) => {
+  try {
+    localStorage.removeItem(key);
+  } catch (e) {
+  }
+};