+++ web/vueapp/src/services/storage/setItem.js
@@ -1,0 +1,6 @@
+export default (key, data) => {
+  try {
+    localStorage.setItem(key, data);
+  } catch (e) {
+  }
+};